backend/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
pydantic
python-dotenv
openai
groq
beautifulsoup4
pypdf2
pdf2image
pillow
requests
lxml

